Moved Facility Screen into its own folder for consistent app structure.
